requirements/base.txt
Django==1.8.4
psycopg2>=2.5.4
pycrypto>=2.6
python-dotenv==0.1.3
